check dsa before flash

+++ host_setup_v2.sh
@@ -8,32 +8,31 @@ usage() {
     echo "Running host_setup.sh to install XRT and flash Shell on host machine. "
     echo ""
     echo "Usage:"
-    echo "  ./host_setup.sh --platform <platform> --version <version>"
-    echo "  ./host_setup.sh  -p <platform>         -v       <version>"
-    echo "  <platform>       : alveo-u200 / alveo-u250 / alveo-u280"
+    echo "  ./host_setup_v2.sh --version <version>"
+    echo "  ./host_setup_v2.sh  -v       <version>"
     echo "  <version>        : 2018.3 / 2019.1 / 2019.2"
     echo "  --skip-xrt-install    : skip install XRT"
     echo "  --skip-shell-flash    : skip flash Shell"
+    echo "  --install-docker      : install docker serveice"
     echo ""
     echo "Example:"
-    echo "Install 2019.2 XRT for Alveo U200 and flash Shell "
-    echo "  ./install.sh -p alveo-u200 -v 2019.2"
+    echo "Install 2019.2 XRT and flash Shell "
+    echo "  ./host_setup_v2.sh -v 2019.2"
 
 }
 
 list() {
-    echo "Available Docker Images:"
     echo ""
-    echo "Image Name                     Support Platform              Version      OS Version"
-    echo "alveo-2018-3-centos            Alveo U200 / U250             2018.3       CentOS"
-    echo "alveo-2018-3-ubuntu-1604       Alveo U200 / U250             2018.3       Ubuntu 16.04"
-    echo "alveo-2018-3-ubuntu-1804       Alveo U200 / U250             2018.3       Ubuntu 18.04"
-    echo "alveo-2019-1-centos            Alveo U200 / U250 / U280      2019.1       CentOS"
-    echo "alveo-2019-1-ubuntu-1604       Alveo U200 / U250 / U280      2019.1       Ubuntu 16.04"
-    echo "alveo-2019-1-ubuntu-1804       Alveo U200 / U250 / U280      2019.1       Ubuntu 18.04"
-    echo "alveo-2019-2-centos            Alveo U200 / U250 / U280      2019.2       CentOS"
-    echo "alveo-2019-2-ubuntu-1604       Alveo U200 / U250 / U280      2019.2       Ubuntu 16.04"
-    echo "alveo-2019-2-ubuntu-1804       Alveo U200 / U250 / U280      2019.2       Ubuntu 18.04"
+    echo "Support Platform              Version      OS Version"
+    echo "Alveo U200 / U250             2018.3       CentOS"
+    echo "Alveo U200 / U250             2018.3       Ubuntu 16.04"
+    echo "Alveo U200 / U250             2018.3       Ubuntu 18.04"
+    echo "Alveo U200 / U250 / U280      2019.1       CentOS"
+    echo "Alveo U200 / U250 / U280      2019.1       Ubuntu 16.04"
+    echo "Alveo U200 / U250 / U280      2019.1       Ubuntu 18.04"
+    echo "Alveo U200 / U250 / U280      2019.2       CentOS"
+    echo "Alveo U200 / U250 / U280      2019.2       Ubuntu 16.04"
+    echo "Alveo U200 / U250 / U280      2019.2       Ubuntu 18.04"
 }
 
 get_packages() {
@@ -44,6 +43,8 @@ get_packages() {
         SHELL_PACKAGE=`grep ^$COMB: conf/spec.txt | awk -F':' '{print $2}' | awk -F';' '{print $2}' | awk -F= '{print $2}'`
         DSA=`grep ^$COMB: conf/spec.txt | awk -F':' '{print $2}' | awk -F';' '{print $3}' | awk -F= '{print $2}'`
         TIMESTAMP=`grep ^$COMB: conf/spec.txt | awk -F':' '{print $2}' | awk -F';' '{print $4}' | awk -F= '{print $2}'`
+        PACKAGE_NAME=`grep ^$COMB: conf/spec.txt | awk -F':' '{print $2}' | awk -F';' '{print $6}' | awk -F= '{print $2}'`
+        PACKAGE_VERSION=`grep ^$COMB: conf/spec.txt | awk -F':' '{print $2}' | awk -F';' '{print $7}' | awk -F= '{print $2}'`
     else
         usage
         echo ""
@@ -59,9 +60,6 @@ install_xrt() {
     echo "Install XRT"
     if [[ "$OSVERSION" == "ubuntu-16.04" ]] || [[ "$OSVERSION" == "ubuntu-18.04" ]]; then
         apt-get install --reinstall /tmp/$XRT_PACKAGE
-    # elif [[ "$OSVERSION" == "redhat" ]]; then
-    #       yum-config-manager --enable rhel-7-server-optional-rpms
-    #       yum install -y https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
     elif [[ "$OSVERSION" == "centos" ]]; then
         yum install epel-release
         yum install /tmp/$XRT_PACKAGE
@@ -69,21 +67,36 @@ install_xrt() {
     rm /tmp/$XRT_PACKAGE
 }
 
+check_packages() {
+    if [[ "$OSVERSION" == "ubuntu-16.04" ]] || [[ "$OSVERSION" == "ubuntu-18.04" ]]; then
+        PACKAGE_INSTALL_INFO=`apt list --installed 2>/dev/null | grep "$PACKAGE_NAME" | grep "$PACKAGE_VERSION"`
+    elif [[ "$OSVERSION" == "centos" ]]; then
+        PACKAGE_INSTALL_INFO=`yum list installed 2>/dev/null | grep "$PACKAGE_NAME" | grep "$PACKAGE_VERSION"`
+    fi
+}
+
 flash_cards() {
     get_packages
-    echo "Download Shell package"
-    wget -cO - "https://www.xilinx.com/bin/public/openDownload?filename=$SHELL_PACKAGE" > /tmp/$SHELL_PACKAGE
+    check_packages
+    if [[ $? != 0 ]]; then
+        echo "Download Shell package"
+        wget -cO - "https://www.xilinx.com/bin/public/openDownload?filename=$SHELL_PACKAGE" > /tmp/$SHELL_PACKAGE
 
-    echo "Install Shell"
-    if [[ "$OSVERSION" == "ubuntu-16.04" ]] || [[ "$OSVERSION" == "ubuntu-18.04" ]]; then
-        apt-get install /tmp/$SHELL_PACKAGE
-    elif [[ "$OSVERSION" == "centos" ]]; then
-        rpm -i /tmp/$SHELL_PACKAGE
+        echo "Install Shell"
+        if [[ "$OSVERSION" == "ubuntu-16.04" ]] || [[ "$OSVERSION" == "ubuntu-18.04" ]]; then
+            apt-get install /tmp/$SHELL_PACKAGE
+        elif [[ "$OSVERSION" == "centos" ]]; then
+            rpm -i /tmp/$SHELL_PACKAGE
+        fi
+        rm /tmp/$SHELL_PACKAGE
     fi
-    rm /tmp/$SHELL_PACKAGE
     
     echo "Flash Card"
-    /opt/xilinx/xrt/bin/xbutil flash -a $DSA  $TIMESTAMP
+    if [[ -f /opt/xilinx/xrt/bin/xbmgmt ]]; then
+        /opt/xilinx/xrt/bin/xbmgmt flash -a $DSA  $TIMESTAMP
+    else
+        /opt/xilinx/xrt/bin/xbutil flash -a $DSA  $TIMESTAMP
+    fi
 }
 
 notice_disclaimer() {
@@ -187,6 +200,7 @@ if [[ "$SHELL" == 1 ]]; then
 
     if [[ "$U200" == 0 && "$U250" == 0 && "$U280" == 0 ]]; then
         echo "[ERROR] No FPGA Board Detected. You need at least one FPGA board to use this script."
+        list
         exit 1;
     fi
